refactored the regex set (re)construction work code into a separate file/module: `regexp-set-management.js` as this is a HUGE chunk of code that's only sideways related to producing a regexp-based lexer: this code is used to make jison lexer **macro expansion** more powerful as it allows you to expand lexer macros inside regexp `[...]` *sets* and thus construct compound sets in match regexes. With https://github.com/GerHobbelt/jison/issues/6 still dangling, this code is only expected to become more complex when we add full 'astral Unicode' support, so factoring it all out into its own separate file/module is paramount.
